+++ src/tko/game/task.py
@@ -21,6 +21,9 @@ class Task(TreeItem):
     # default values for tasks
     approach_max = 6
     autonomy_max = 5
+    description_max = 5
+    desire_max = 5
+    effort_max = 5
 
     def __init__(self):
 
@@ -32,6 +35,10 @@ class Task(TreeItem):
         self.approach: int = 0 # valor de 0 a approach_max
         self.autonomy: int = 0 # valor de 0 a autonomy_max
         self.main_idx: int = 0
+
+        self.description: int = 0
+        self.desire: int = 0
+        self.effort: int = 0
 
         self.qskills: dict[str, int] = {} # default quest skills
         self.skills: dict[str, int] = {} # local skills
@@ -85,6 +92,12 @@ class Task(TreeItem):
                     self.approach = int(val)
                 elif k == "hab" or k == "auto":
                     self.autonomy = int(val)
+                elif k == "desc":
+                    self.description = int(val)
+                elif k == "desire":
+                    self.desire = int(val)
+                elif k == "effort":
+                    self.effort = int(val)
                 elif k == "idx":
                     self.main_idx = int(val)
         elif ":" not in value:
@@ -102,7 +115,7 @@ class Task(TreeItem):
                 self.main_idx = (int(v[3]))
 
     def save_to_db(self) -> str:
-        return "{" + f"cov:{self.coverage}, app:{self.approach}, auto:{self.autonomy}, idx:{self.main_idx}" + "}"
+        return "{" + f"cov:{self.coverage}, appr:{self.approach}, auto:{self.autonomy}, desc:{self.description}, desire:{self.desire}, effort:{self.effort}, idx:{self.main_idx}" + "}"
     
     def is_db_empty(self) -> bool:
         return self.approach == 0 and self.autonomy == 0 and self.main_idx == 0 and self.coverage == 0
@@ -177,6 +190,27 @@ class Task(TreeItem):
             self.autonomy = value
         else:
             print(f"Compreensão inválida: {value}")
+
+    def set_description(self, value: int):
+        value = int(value)
+        if value >= 0 and value <= Task.description_max:
+            self.description = value
+        else:
+            print(f"Descrição inválida: {value}")
+    
+    def set_desire(self, value: int):
+        value = int(value)
+        if value >= 0 and value <= Task.desire_max:
+            self.desire = value
+        else:
+            print(f"Interesse inválido: {value}")
+
+    def set_effort(self, value: int):
+        value = int(value)
+        if value >= 0 and value <= Task.effort_max:
+            self.effort = value
+        else:
+            print(f"Dedicação inválida: {value}")
 
     # @override
     def __str__(self):